Moves universal OCI CDI extension to new location
