+++ docs/mp/extensions/04_cdi_oci-objectstorage.adoc
@@ -36,8 +36,8 @@ include::{common-deps-page-prefix-inc}[tag=maven-dependency]
 [source,xml]
 ----
 <dependency>
-     <groupId>io.helidon.integrations.cdi</groupId>
-     <artifactId>helidon-integrations-cdi-oci</artifactId>
+     <groupId>io.helidon.integrations.oci.sdk</groupId>
+     <artifactId>helidon-integrations-oci-sdk-cdi</artifactId>
 </dependency>
 ----
