+++ integrations/oci/pom.xml
@@ -42,5 +42,6 @@
         <module>cdi</module>
         <module>atp</module>
         <module>metrics</module>
+        <module>sdk</module>
     </modules>
 </project>
