+++ integrations/oci/sdk/cdi/src/test/resources/logging.properties
@@ -17,7 +17,7 @@
 
 handlers = java.util.logging.ConsoleHandler
 
-io.helidon.integrations.cdi.oci.level = FINE
+io.helidon.integrations.oci.sdk.cdi.level = FINE
 
 java.util.logging.ConsoleHandler.level = ALL
 java.util.logging.ConsoleHandler.formatter = java.util.logging.SimpleFormatter
